+++ Nesting_4/IUtilities.cs
@@ -8,7 +8,7 @@ namespace Nesting_4
 {
     interface IUtilities
     {
-        float ComputeLowerBound(IList<Item> pricedItems, int binWidth, int binHeight);
+        double ComputeLowerBound(IList<Item> pricedItems, int binWidth, int binHeight);
 
         Bin<Tuple> IsBestPositionFound(Bin<Tuple> temporaryBin, Item temporaryPricedItem, string itemAllocationMethod);
